+++ analytics/src/main/java/no/nordicsemi/android/common/analytics/AnalyticsPermissionData.kt
@@ -29,29 +29,9 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.ui
+package no.nordicsemi.android.common.analytics
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.text.buildAnnotatedString
-import kotlinx.coroutines.flow.MutableStateFlow
-import no.nordicsemi.android.common.navigation.ParcelableArgument
-import no.nordicsemi.android.common.navigation.SuccessDestinationResult
-import no.nordicsemi.android.common.ui.scanner.DiscoveredBluetoothDevice
-
-@Composable
-fun String.toAnnotatedString() = buildAnnotatedString {
-    append(this@toAnnotatedString)
-}
-
-fun <T> MutableStateFlow<T>.updateIfDifferent(newValue: T) {
-    if (value != newValue) {
-        value = newValue
-    }
-}
-
-val <T> T.exhaustive: Any
-    get() = this as Any
-
-fun SuccessDestinationResult.getDevice(): DiscoveredBluetoothDevice {
-    return (argument as ParcelableArgument).value as DiscoveredBluetoothDevice
-}
+data class AnalyticsPermissionData(
+    val isPermissionGranted: Boolean = false,
+    val wasInfoDialogShown: Boolean = false
+)
