+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/main/FilterView.kt
@@ -29,7 +29,7 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.view
+package no.nordicsemi.android.common.ui.scanner.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
