+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/main/ScanEmptyView.kt
@@ -29,15 +29,13 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.view.error
+package no.nordicsemi.android.common.ui.scanner.main
 
-import android.bluetooth.BluetoothAdapter
 import android.content.Context
 import android.content.Intent
+import android.provider.Settings
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -47,63 +45,52 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import no.nordicsemi.android.common.theme.parseBold
 import no.nordicsemi.android.common.ui.scanner.R
-import no.nordicsemi.android.common.ui.scanner.ui.AppBar
-import no.nordicsemi.android.common.ui.scanner.view.event.Event
 
 @Composable
-internal fun BluetoothDisabledView(onEvent: (Event) -> Unit) {
-    Column {
-        AppBar(stringResource(id = R.string.scanner_error)) { onEvent(Event.NavigateUp) }
+internal fun ScanEmptyView(requireLocation: Boolean) {
+    Column(
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = Modifier.padding(32.dp)
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_bluetooth_disabled),
+            contentDescription = "",
+            modifier = Modifier.padding(16.dp)
+        )
 
-        Column(
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier.fillMaxSize().verticalScroll(rememberScrollState())
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_bluetooth_disabled),
-                contentDescription = "",
-                modifier = Modifier.padding(16.dp)
-            )
+        Text(
+            text = stringResource(id = R.string.no_device_guide_title),
+            color = MaterialTheme.colorScheme.secondary
+        )
 
-            Text(
-                text = stringResource(id = R.string.bluetooth_disabled_title),
-                color = MaterialTheme.colorScheme.secondary
-            )
+        Spacer(modifier = Modifier.height(16.dp))
 
-            Spacer(modifier = Modifier.height(16.dp))
+        Text(text = stringResource(id = R.string.no_device_guide_info).parseBold())
 
-            Text(
-                text = stringResource(id = R.string.bluetooth_disabled_info),
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                textAlign = TextAlign.Center
-            )
+        if (requireLocation) {
+            Spacer(modifier = Modifier.height(8.dp))
+
+            Text(text = stringResource(id = R.string.no_device_guide_location_info))
 
             Spacer(modifier = Modifier.height(16.dp))
 
             val context = LocalContext.current
-            Button(onClick = { enableBluetooth(context) }) {
-                Text(text = stringResource(id = R.string.action_enable))
+            Button(onClick = { openLocationSettings(context) }) {
+                Text(text = stringResource(id = R.string.action_location_settings))
             }
-
+            Spacer(modifier = Modifier.height(8.dp))
+        } else {
             Spacer(modifier = Modifier.height(16.dp))
         }
     }
 }
 
-private fun enableBluetooth(context: Context) {
-    context.startActivity(Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE))
-}
-
-@Preview
-@Composable
-private fun BluetoothDisabledView_Preview() {
-    BluetoothDisabledView { }
+private fun openLocationSettings(context: Context) {
+    val intent = Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS)
+    intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+    context.startActivity(intent)
 }
