+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/main/error/BluetoothNotAvailableView.kt
@@ -29,66 +29,63 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.view
+package no.nordicsemi.android.common.ui.scanner.main.error
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import no.nordicsemi.android.common.ui.scanner.R
-import no.nordicsemi.android.common.ui.scanner.DiscoveredBluetoothDevice
+import no.nordicsemi.android.common.ui.scanner.main.ScannerAppBar
+import no.nordicsemi.android.common.ui.scanner.navigation.ScannerNavigationEvent
 
 @Composable
-fun DeviceItem(
-    device: DiscoveredBluetoothDevice,
-    modifier: Modifier = Modifier,
-    extras: @Composable (DiscoveredBluetoothDevice) -> Unit = {},
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically)
-    {
-        Image(
-            painter = painterResource(R.drawable.ic_bluetooth),
-            contentDescription = "Content image",
-            colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.onSecondary),
-            modifier = Modifier
-                .background(
-                    color = MaterialTheme.colorScheme.secondary,
-                    shape = CircleShape
-                )
-                .padding(8.dp)
-        )
-
-        Spacer(modifier = Modifier.width(16.dp))
+internal fun BluetoothNotAvailableView(onEvent: (ScannerNavigationEvent) -> Unit) {
+    Column {
+        ScannerAppBar(stringResource(id = R.string.scanner_error)) { onEvent(ScannerNavigationEvent.NavigateUp) }
 
         Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally,
+            modifier = Modifier.fillMaxSize().verticalScroll(rememberScrollState())
         ) {
-            device.displayName?.let { name ->
-                Text(
-                    text = name,
-                    style = MaterialTheme.typography.titleMedium
-                )
-            } ?: Text(
-                text = stringResource(id = R.string.device_no_name),
-                style = MaterialTheme.typography.titleMedium,
-                color = MaterialTheme.colorScheme.outline
+            Image(
+                painter = painterResource(id = R.drawable.ic_bluetooth_disabled),
+                contentDescription = "",
+                modifier = Modifier.padding(16.dp)
             )
-            Text(text = device.address, style = MaterialTheme.typography.bodyMedium)
-        }
 
-        extras(device)
+            Text(
+                text = stringResource(id = R.string.bluetooth_not_available_title),
+                color = MaterialTheme.colorScheme.secondary,
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Text(
+                text = stringResource(id = R.string.bluetooth_not_available_info),
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.fillMaxWidth().padding(16.dp),
+                textAlign = TextAlign.Center
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+        }
     }
+}
+
+@Preview
+@Composable
+private fun BluetoothNotAvailableView_Preview() {
+    BluetoothNotAvailableView { }
 }
