+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/view/DeviceConnectingView.kt
@@ -29,13 +29,17 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.ui
+package no.nordicsemi.android.common.ui.scanner.view
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.HighlightOff
+import androidx.compose.material.icons.filled.HourglassTop
 import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -50,21 +54,15 @@ import androidx.compose.ui.unit.dp
 import no.nordicsemi.android.common.theme.ScreenSection
 import no.nordicsemi.android.common.ui.scanner.R
 
-enum class Reason {
-    USER, UNKNOWN, LINK_LOSS, MISSING_SERVICE
-}
-
 @Composable
-fun DeviceDisconnectedView(reason: Reason, navigateUp: () -> Unit) {
+fun DeviceConnectingView(navigateUp: () -> Unit) {
     Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
+        modifier = Modifier.fillMaxSize().padding(16.dp),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
         ScreenSection {
             Icon(
-                imageVector = Icons.Default.HighlightOff,
+                imageVector = Icons.Default.HourglassTop,
                 contentDescription = null,
                 tint = MaterialTheme.colorScheme.onSecondary,
                 modifier = Modifier
@@ -78,36 +76,37 @@ fun DeviceDisconnectedView(reason: Reason, navigateUp: () -> Unit) {
             Spacer(modifier = Modifier.size(16.dp))
 
             Text(
-                text = stringResource(id = R.string.device_disconnected),
+                text = stringResource(id = R.string.device_connecting),
                 style = MaterialTheme.typography.titleMedium
             )
 
             Spacer(modifier = Modifier.size(16.dp))
 
-            val text = when (reason) {
-                Reason.USER -> stringResource(id = R.string.device_reason_user)
-                Reason.LINK_LOSS -> stringResource(id = R.string.device_reason_link_loss)
-                Reason.MISSING_SERVICE -> stringResource(id = R.string.device_reason_missing_service)
-                Reason.UNKNOWN -> stringResource(id = R.string.device_reason_unknown)
-            }
-
             Text(
-                text = text,
+                text = stringResource(id = R.string.device_explanation),
                 textAlign = TextAlign.Center,
                 style = MaterialTheme.typography.bodyMedium
+            )
+
+            Spacer(modifier = Modifier.size(16.dp))
+
+            Text(
+                text = stringResource(id = R.string.device_please_wait),
+                textAlign = TextAlign.Center,
+                style = MaterialTheme.typography.titleLarge
             )
         }
 
         Spacer(modifier = Modifier.size(16.dp))
 
         Button(onClick = { navigateUp() }) {
-            Text(text = stringResource(id = R.string.go_up))
+            Text(text = stringResource(id = R.string.cancel))
         }
     }
 }
 
 @Preview
 @Composable
-fun DeviceDisconnectedView_Preview() {
+fun DeviceConnectingView_Preview() {
     DeviceConnectingView { }
 }
